move `ElevatesSessions` into `Statamic\Testing\Concerns`
